Add database and networking modules
